backtest_engine/requirements.txt
# requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.2

# Database
supabase>=2.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Data fetching
polygon-api-client>=1.12.0
requests>=2.31.0
aiohttp>=3.8.0

# UI (PyQt6)
PyQt6>=6.5.0
PyQt6-Charts>=6.5.0
pyqtgraph>=0.13.0

# Analysis & Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Performance & Optimization
numba>=0.57.0
joblib>=1.3.0
multiprocess>=0.70.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Development tools
black>=23.7.0
pylint>=2.17.0
mypy>=1.4.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0
loguru>=0.7.0

# Reports
reportlab>=4.0.0
jinja2>=3.1.0
weasyprint>=59.0  # For HTML to PDF conversion